Adding an install script.
This is really close.  The output compares well with the purely fortran
version for aerosols and gases.  The output file can be read by IOAPI
and is recognized as a METBDY file.  There is still some work that
needs to be done setting SDATE, STIME, EDATE, and ETIME after the file
is initialized.